reduce duplication across Vi and Emacs editors
