+++ src/edit_mode/emacs.rs
@@ -7,10 +7,11 @@ use crate::{
         },
         EditMode,
     },
-    enums::{EditCommand, ReedlineEvent, ReedlineRawEvent},
+    enums::{EditCommand, ReedlineEvent},
     PromptEditMode,
 };
-use crossterm::event::{Event, KeyCode, KeyEvent, KeyModifiers};
+use crossterm::event::{KeyCode, KeyModifiers};
+
 
 /// Returns the current default emacs keybindings
 pub fn default_emacs_keybindings() -> Keybindings {
@@ -119,33 +120,13 @@ impl Default for Emacs {
 }
 
 impl EditMode for Emacs {
-    fn parse_event(&mut self, event: ReedlineRawEvent) -> ReedlineEvent {
-        match event.into() {
-            Event::Key(KeyEvent {
-                code, modifiers, ..
-            }) => {
-                let combo = Self::normalize_key_combo(modifiers, code);
-                let keybindings = &self.keybindings;
-                let resolution = self
-                    .sequence_state
-                    .process_combo(keybindings, combo);
-                resolution
-                    .into_event(|combo| Self::single_key_event(keybindings, combo))
-                    .unwrap_or(ReedlineEvent::None)
-            }
-
-            Event::Mouse(_) => self.with_flushed_sequence(ReedlineEvent::Mouse),
-            Event::Resize(width, height) => {
-                self.with_flushed_sequence(ReedlineEvent::Resize(width, height))
-            }
-            Event::FocusGained => self.with_flushed_sequence(ReedlineEvent::None),
-            Event::FocusLost => self.with_flushed_sequence(ReedlineEvent::None),
-            Event::Paste(body) => {
-                self.with_flushed_sequence(ReedlineEvent::Edit(vec![EditCommand::InsertString(
-                    body.replace("\r\n", "\n").replace('\r', "\n"),
-                )]))
-            }
-        }
+    fn parse_key_event(&mut self, modifiers: KeyModifiers, code: KeyCode) -> ReedlineEvent {
+        let combo = KeyCombination::from((modifiers, code));
+        let keybindings = &self.keybindings;
+        let resolution = self.sequence_state.process_combo(keybindings, combo);
+        resolution
+            .into_event(|combo| self.default_key_event(keybindings, combo))
+            .unwrap_or(ReedlineEvent::None)
     }
 
     fn edit_mode(&self) -> PromptEditMode {
@@ -159,7 +140,7 @@ impl EditMode for Emacs {
     fn flush_pending_sequence(&mut self) -> Option<ReedlineEvent> {
         let keybindings = &self.keybindings;
         let resolution = self.sequence_state.flush_with_combos();
-        resolution.into_event(|combo| Self::single_key_event(keybindings, combo))
+        resolution.into_event(|combo| self.default_key_event(keybindings, combo))
     }
 }
 
@@ -171,72 +152,13 @@ impl Emacs {
             sequence_state: KeySequenceState::new(),
         }
     }
-
-    fn normalize_key_combo(modifier: KeyModifiers, code: KeyCode) -> KeyCombination {
-        let key_code = match code {
-            KeyCode::Char(c) => {
-                let c = match modifier {
-                    KeyModifiers::NONE => c,
-                    _ => c.to_ascii_lowercase(),
-                };
-                KeyCode::Char(c)
-            }
-            other => other,
-        };
-
-        KeyCombination { modifier, key_code }
-    }
-
-    fn single_key_event(keybindings: &Keybindings, combo: KeyCombination) -> ReedlineEvent {
-        match combo.key_code {
-            KeyCode::Char(c) => keybindings
-                .find_binding(combo.modifier, KeyCode::Char(c))
-                .unwrap_or_else(|| {
-                    if combo.modifier == KeyModifiers::NONE
-                        || combo.modifier == KeyModifiers::SHIFT
-                        || combo.modifier == KeyModifiers::CONTROL | KeyModifiers::ALT
-                        || combo.modifier
-                            == KeyModifiers::CONTROL | KeyModifiers::ALT | KeyModifiers::SHIFT
-                    {
-                        ReedlineEvent::Edit(vec![EditCommand::InsertChar(
-                            if combo.modifier == KeyModifiers::SHIFT {
-                                c.to_ascii_uppercase()
-                            } else {
-                                c
-                            },
-                        )])
-                    } else {
-                        ReedlineEvent::None
-                    }
-                }),
-            code => keybindings
-                .find_binding(combo.modifier, code)
-                .unwrap_or(ReedlineEvent::None),
-        }
-    }
-
-    fn with_flushed_sequence(&mut self, event: ReedlineEvent) -> ReedlineEvent {
-        let Some(flush_event) = self.flush_pending_sequence() else {
-            return event;
-        };
-
-        if matches!(event, ReedlineEvent::None) {
-            return flush_event;
-        }
-
-        match flush_event {
-            ReedlineEvent::Multiple(mut events) => {
-                events.push(event);
-                ReedlineEvent::Multiple(events)
-            }
-            other => ReedlineEvent::Multiple(vec![other, event]),
-        }
-    }
 }
 
 #[cfg(test)]
 mod test {
     use super::*;
+    use crate::enums::ReedlineRawEvent;
+    use crossterm::event::{Event, KeyEvent};
     use pretty_assertions::assert_eq;
 
     #[test]
